Split up into smaller modules
- moved helper functions out of main fractional_indexing.py
- updated tests.py with new imports and error name
