+++ exceptions.py
@@ -1,0 +1,5 @@
+# fractional_indexing/exceptions.py
+
+class OrderKeyError(Exception):
+    """Custom error for invalid order keys."""
+    pass
