+++ tests.py
@@ -2,8 +2,12 @@ from typing import Optional
 
 import pytest
 
-from fractional_indexing import FIError, generate_key_between, generate_n_keys_between, validate_order_key
-
+from fractional_indexing import (
+    OrderKeyError,
+    generate_key_between,
+    generate_n_keys_between,
+)
+from .utils import validate_order_key
 
 BASE_95_DIGITS = ' !"#$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_`abcdefghijklmnopqrstuvwxyz{|}~'
 
@@ -28,18 +32,18 @@ BASE_95_DIGITS = ' !"#$%&\'()*+,-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\
     ('Zz', 'a01', 'a0'),
     (None, 'a0V', 'a0'),
     (None, 'b999', 'b99'),
-    (None, 'A00000000000000000000000000', FIError('invalid order key: A00000000000000000000000000')),
+    (None, 'A00000000000000000000000000', OrderKeyError('invalid order key: A00000000000000000000000000')),
     (None, 'A000000000000000000000000001', 'A000000000000000000000000000V'),
     ('zzzzzzzzzzzzzzzzzzzzzzzzzzy', None, 'zzzzzzzzzzzzzzzzzzzzzzzzzzz'),
     ('zzzzzzzzzzzzzzzzzzzzzzzzzzz', None, 'zzzzzzzzzzzzzzzzzzzzzzzzzzzV'),
-    ('a00', None, FIError('invalid order key: a00')),
-    ('a00', 'a1', FIError('invalid order key: a00')),
-    ('0', '1', FIError('invalid order key head: 0')),
-    ('a1', 'a0', FIError('a1 >= a0')),
+    ('a00', None, OrderKeyError('invalid order key: a00')),
+    ('a00', 'a1', OrderKeyError('invalid order key: a00')),
+    ('0', '1', OrderKeyError('invalid order key head: 0')),
+    ('a1', 'a0', OrderKeyError('a1 >= a0')),
 ])
-def test_generate_key_between(a: Optional[str], b: Optional[str], expected: str) -> None:
-    if isinstance(expected, FIError):
-        with pytest.raises(FIError) as e:
+def test_generate_key_between(a: Optional[str], b: Optional[str], expected) -> None:
+    if isinstance(expected, OrderKeyError):
+        with pytest.raises(OrderKeyError) as e:
             generate_key_between(a, b)
         assert e.value.args[0] == expected.args[0]
     else:
@@ -72,23 +76,23 @@ def test_generate_n_keys_between(a: Optional[str], b: Optional[str], n: int, exp
     (None, None, 'a '),
     ('a ', None, 'a!'),
     (None, 'a ', 'Z~'),
-    ('a0 ', 'a0!', FIError('invalid order key: a0 ')),
+    ('a0 ', 'a0!', OrderKeyError('invalid order key: a0 ')),
     (None, 'A                          0', 'A                          ('),
     ('a~', None, 'b  '),
     ('Z~', None, 'a '),
-    ('b   ', None, FIError('invalid order key: b   ')),
+    ('b   ', None, OrderKeyError('invalid order key: b   ')),
     ('a0', 'a0V', 'a0;'),
     ('a  1', 'a  2', 'a  1P'),
-    (None, 'A                          ', FIError('invalid order key: A                          ')),
+    (None, 'A                          ', OrderKeyError('invalid order key: A                          ')),
 ])
 def test_base95_digits(a: Optional[str], b: Optional[str], expected: str) -> None:
     kwargs = {
-        'a': a,
-        'b': b,
+        'start_key': a,
+        'end_key': b,
         'digits': BASE_95_DIGITS,
     }
-    if isinstance(expected, FIError):
-        with pytest.raises(FIError) as e:
+    if isinstance(expected, OrderKeyError):
+        with pytest.raises(OrderKeyError) as e:
             generate_key_between(**kwargs)
         assert e.value.args[0] == expected.args[0]
     else:
@@ -124,31 +128,30 @@ def test_readme_examples_single_key():
 
 def test_readme_examples_multiple_keys():
     # Insert 3 at the beginning
-    keys = generate_n_keys_between(None, None, n=3)
+    keys = generate_n_keys_between(None, None, number_of_keys=3)
     assert keys == ['a0', 'a1', 'a2']
 
     # Insert 3 after 1st
-    keys = generate_n_keys_between('a0', None, n=3)
+    keys = generate_n_keys_between('a0', None, number_of_keys=3)
     assert keys == ['a1', 'a2', 'a3']
 
     # Insert 3 before 1st
-    keys = generate_n_keys_between(None, 'a0', n=3)
+    keys = generate_n_keys_between(None, 'a0', number_of_keys=3)
     assert keys == ['Zx', 'Zy', 'Zz']
 
     # Insert 3 in between 2nd and 3rd. Midpoint
-    keys = generate_n_keys_between('a1', 'a2', n=3)
+    keys = generate_n_keys_between('a1', 'a2', number_of_keys=3)
     assert keys == ['a1G', 'a1V', 'a1l']
 
 
 def test_readme_examples_validate_order_key():
-    from fractional_indexing import validate_order_key, FIError
+    from fractional_indexing import validate_order_key, OrderKeyError
 
     validate_order_key('a0')
 
-    try:
+    with pytest.raises(OrderKeyError) as e:
         validate_order_key('foo')
-    except FIError as e:
-        print(e)  # fractional_indexing.FIError: invalid order key: foo
+    assert str(e.value) == 'invalid order key: foo'
 
 
 def test_readme_examples_custom_base():
@@ -156,4 +159,4 @@ def test_readme_examples_custom_base():
     assert generate_key_between(None, None, digits=BASE_95_DIGITS) == 'a '
     assert generate_key_between('a ', None, digits=BASE_95_DIGITS) == 'a!'
     assert generate_key_between(None, 'a ', digits=BASE_95_DIGITS) == 'Z~'
-    assert generate_n_keys_between('a ', 'a!', n=3, digits=BASE_95_DIGITS) == ['a 8', 'a P', 'a h']
+    assert generate_n_keys_between('a ', 'a!', number_of_keys=3, digits=BASE_95_DIGITS) == ['a 8', 'a P', 'a h']
